requirements.txt
Flask==1.1.1
pymongo==3.10.1
dnspython==1.16.0
beautifulsoup4==4.8.2
requests==2.23.0